Moving tower building to the sliding window algorithm
